Move FormTracker to its own module
