+++ src/form-tracker.js
@@ -1,0 +1,25 @@
+import {_} from './utils';
+import DomTracker from './dom-tracker';
+
+/**
+ * FormTracker Object
+ * @constructor
+ * @extends DomTracker
+ */
+var FormTracker = function() {
+    this.override_event = 'submit';
+};
+_.inherit(FormTracker, DomTracker);
+
+FormTracker.prototype.event_handler = function(evt, element, options) {
+    options.element = element;
+    evt.preventDefault();
+};
+
+FormTracker.prototype.after_track_handler = function(props, options) {
+    setTimeout(function() {
+        options.element.submit();
+    }, 0);
+};
+
+export default FormTracker;
